1.12 -> feat : add carousel tap animation,add popup Info system

+++ src/view/iGameModel/IGameModelDebug.java
@@ -4,5 +4,5 @@ public interface IGameModelDebug {
     int getIce_BasicCount();
     int getIce_RareCount();
     int getIce_LegendaryCount();
-    long getPlayTick();
+    int getPlayTick();
 }
